refactor: apply Vercel React best practices — 9 performance and pattern improvements
- Remove duplicate cache() wrappers (fetchPostById already cached at action level)
- Replace && with ternary conditionals in JSX to prevent falsy value rendering
- Use Promise.all for parallel data fetching in posts/[slug] layout
- Add Suspense boundaries with Loader fallback on all content pages
- Split types/index.ts barrel file into individual modules for tree-shaking
- Extract HomeContent to separate file to prevent remounting on parent re-renders
- Stabilize MusicPlayer callbacks with isPlayingRef (stable references, no context re-renders)
- Dynamic import Comments (giscus) with ssr:false to defer heavy third-party bundle
- Extract shared useVideoWithPlaceholder hook from MediaBackground and VideoWithPlaceholder
