+++ app/about/page.tsx
@@ -9,8 +9,10 @@ import { Metadata } from "next";
 import { siteMetadata } from "@/site/siteMetadata";
 // import { ogNoteImage } from "@/site/data";
 
+import { Suspense } from "react";
 import AboutPage from "@/app/about/components/AboutPage";
 import Content from "@/components/contents/Content";
+import Loader from "@/components/common/Loader";
 import { fetchStaticPageContent } from "@/actions/posts";
 
 export const metadata: Metadata = {
@@ -25,7 +27,11 @@ const About = async () => {
   const recordMap = await fetchStaticPageContent("about");
 
   return (
-    <AboutPage>{recordMap && <Content recordMap={recordMap} />}</AboutPage>
+    <AboutPage>
+      <Suspense fallback={<Loader />}>
+        {recordMap ? <Content recordMap={recordMap} /> : null}
+      </Suspense>
+    </AboutPage>
   );
 };
