+++ app/home/HomeContent.tsx
@@ -1,0 +1,34 @@
+"use client";
+
+import { useEffect } from "react";
+import { useHotkeys } from "@mantine/hooks";
+import HomePage from "@/components/home/HomePage";
+import NewProfile from "@/components/home/NewProfile";
+import { useModeStore } from "@/hooks/use-mode-store";
+import { useMusicPlayer } from "@/context/MusicPlayerContext";
+
+const HomeContent = () => {
+  const mode = useModeStore((state) => state.mode);
+  const setMode = useModeStore((state) => state.setMode);
+  const { play, pause } = useMusicPlayer();
+
+  useEffect(() => {
+    useModeStore.persist.rehydrate();
+  }, []);
+
+  useEffect(() => {
+    if (mode === "jianghu") {
+      play();
+    } else {
+      pause();
+    }
+  }, [mode, play, pause]);
+
+  useHotkeys([
+    ["ctrl+Q", () => setMode(mode === "focused" ? "jianghu" : "focused")],
+  ]);
+
+  return <>{mode === "focused" ? <HomePage /> : <NewProfile />}</>;
+};
+
+export default HomeContent;
