+++ app/note/page.tsx
@@ -9,7 +9,9 @@ import { Metadata } from "next";
 import { siteMetadata } from "@/site/siteMetadata";
 // import { ogNoteImage } from "@/site/data";
 
+import { Suspense } from "react";
 import Content from "@/components/contents/Content";
+import Loader from "@/components/common/Loader";
 import { fetchStaticPageContent } from "@/actions/posts";
 import NotePage from "./components/NotePage";
 
@@ -24,7 +26,13 @@ export const metadata: Metadata = {
 const Note = async () => {
   const recordMap = await fetchStaticPageContent("note");
 
-  return <NotePage>{recordMap && <Content recordMap={recordMap} />}</NotePage>;
+  return (
+    <NotePage>
+      <Suspense fallback={<Loader />}>
+        {recordMap ? <Content recordMap={recordMap} /> : null}
+      </Suspense>
+    </NotePage>
+  );
 };
 
 export default Note;
