+++ components/contents/ContentBody.tsx
@@ -1,12 +1,17 @@
 "use client";
 
 import Link from "next/link";
+import dynamic from "next/dynamic";
 import { Suspense } from "react";
 import { Space, Box, Divider } from "@mantine/core";
 import { PageDataSchemaType } from "@/types";
 import Loader from "../common/Loader";
 import RelatedPostsWrapper from "./RelatedPostsWrapper";
-import Comments from "./Comments";
+
+const Comments = dynamic(() => import("./Comments"), {
+  ssr: false,
+  loading: () => <div className="h-32" />,
+});
 
 type Props = {
   postData: PageDataSchemaType;
