+++ components/ui/MediaBackground.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { useEffect, useRef, useState } from "react";
 import Image from "next/image";
 import { cn } from "@/lib/utils";
+import useVideoWithPlaceholder from "@/hooks/use-video-with-placeholder";
 
 interface MediaBackgroundProps {
   children?: React.ReactNode;
@@ -36,40 +36,7 @@ const VideoBackground = ({
   placeholder,
   className,
 }: VideoBackgroundProps) => {
-  const videoRef = useRef<HTMLVideoElement>(null);
-  const [videoLoaded, setVideoLoaded] = useState(false);
-
-  useEffect(() => {
-    if (process.env.NODE_ENV === "development" && !placeholder) {
-      console.warn("No placeholder provided for video");
-    }
-  }, [placeholder]);
-
-  useEffect(() => {
-    const video = videoRef.current;
-    if (!video) return;
-
-    const handleVideoReady = () => setVideoLoaded(true);
-
-    video.addEventListener("loadeddata", handleVideoReady);
-    video.addEventListener("canplay", handleVideoReady);
-    video.load();
-
-    if (video.readyState >= 2) {
-      setVideoLoaded(true);
-    }
-
-    return () => {
-      video.removeEventListener("loadeddata", handleVideoReady);
-      video.removeEventListener("canplay", handleVideoReady);
-    };
-  }, [src]);
-
-  useEffect(() => {
-    if (videoRef.current && videoLoaded) {
-      videoRef.current.play();
-    }
-  }, [videoLoaded]);
+  const { videoRef, videoLoaded } = useVideoWithPlaceholder({ src, placeholder });
 
   return (
     <>
